Fix all CI failures: mypy, ruff, bandit, smoke test, and STT test loop
- Fix 14 mypy errors: rename reused callback variables in engine.py,
  add explicit ctypes casts in speex/rnnoise, fix bytes->bytearray type
  in local backend, type-ignore fastrtc optional dep, rename shadowed
  turn event variable in voice.py
- Fix ruff I001: auto-sort imports in test_pipeline_engine.py
- Fix bandit B108/B110: add nosec comments for mock tmp path and
  intentional except-pass in local_transport cleanup
- Fix smoke test: replace --system pip install with uv venv (PEP 668)
- Fix STT test infinite loop: test_batch_transcribe_transducer used
  OfflineRecognizer mock but transcribe() now uses OnlineRecognizer
  for transducer mode — mock is_ready was truthy MagicMock causing
  infinite while loop. Updated to use OnlineRecognizer mock with
  is_ready side_effect=[True, False]
- Update whisper test to assert decode_stream (not decode)
- Update streaming test to return strings from get_result (not objects)

+++ src/roomkit/voice/pipeline/vad/energy.py
@@ -24,7 +24,7 @@ def _rms_int16(data: bytes) -> float:
         return 0.0
     samples = struct.unpack(f"<{n_samples}h", data[: n_samples * 2])
     sum_sq = sum(s * s for s in samples)
-    return (sum_sq / n_samples) ** 0.5
+    return float((sum_sq / n_samples) ** 0.5)
 
 
 class EnergyVADProvider(VADProvider):
